removed test file and added code to existing file
